+++ frontend/code/components/dropdown.ts
@@ -8,10 +8,6 @@ import {
     KeyboardFocusableComponentState,
 } from "./keyboardFocusableComponent";
 
-const SELECT_OPTION_EVENT = DropdownPositioner.USE_MOBILE_MODE
-    ? "click"
-    : "pointerdown";
-
 export type DropdownState = KeyboardFocusableComponentState & {
     _type_: "Dropdown-builtin";
     optionNames: string[];
@@ -31,6 +27,7 @@ export class DropdownComponent extends KeyboardFocusableComponent<DropdownState>
     private mobileLabelElement: HTMLElement;
 
     private popupManager: PopupManager;
+    private useMobilePopup: boolean;
 
     // The currently highlighted option, if any
     private highlightedOptionElement: HTMLElement | null = null;
@@ -50,11 +47,6 @@ export class DropdownComponent extends KeyboardFocusableComponent<DropdownState>
             connectClickHandlers: false,
         });
         element.appendChild(this.inputBox.outerElement);
-
-        // On mobile, make the input box read-only so that the keyboard doesn't
-        // appear when the Dropdown is focused
-        this.inputBox.inputElement.readOnly =
-            DropdownPositioner.USE_MOBILE_MODE;
 
         // In order to ensure the dropdown can actually fit its options, add a
         // hidden element that will contain a copy of all options. This element
@@ -116,12 +108,20 @@ export class DropdownComponent extends KeyboardFocusableComponent<DropdownState>
             anchor: element,
             content: this.popupElement,
             positioner: new DropdownPositioner(),
-            modal: DropdownPositioner.USE_MOBILE_MODE,
+            modal: false,
             userClosable: true,
             onUserClose: this.hidePopupDontCommit.bind(this),
         });
 
         return element;
+    }
+
+    /// With a `click` handler, the <input> element loses focus for a little
+    /// while, which is noticeable because the floating label will quickly move
+    /// down and then back up. To avoid this, we use `pointerdown` instead. But
+    /// on mobile this prevents scrolling, so mobile uses `click`.
+    private getOptionSelectEvent(): string {
+        return this.useMobilePopup ? "click" : "pointerdown";
     }
 
     /// Open the dropdown and show all options
@@ -181,11 +181,14 @@ export class DropdownComponent extends KeyboardFocusableComponent<DropdownState>
         // Enter: select the highlighted option
         else if (event.key === "Enter") {
             if (this.highlightedOptionElement !== null) {
-                let pointerDownEvent = new PointerEvent(SELECT_OPTION_EVENT, {
-                    bubbles: true,
-                    cancelable: true,
-                    view: window,
-                });
+                let pointerDownEvent = new PointerEvent(
+                    this.getOptionSelectEvent(),
+                    {
+                        bubbles: true,
+                        cancelable: true,
+                        view: window,
+                    }
+                );
 
                 this.highlightedOptionElement.dispatchEvent(pointerDownEvent);
             }
@@ -193,7 +196,7 @@ export class DropdownComponent extends KeyboardFocusableComponent<DropdownState>
 
         // Move highlight down
         else if (event.key === "ArrowDown") {
-            let nextOption;
+            let nextOption: Element | null;
 
             if (this.highlightedOptionElement === null) {
                 nextOption = this.popupOptionsElement.firstElementChild;
@@ -205,7 +208,7 @@ export class DropdownComponent extends KeyboardFocusableComponent<DropdownState>
                 }
             }
 
-            this._highlightOption(nextOption as HTMLElement);
+            this._highlightOption(nextOption as HTMLElement | null);
         }
 
         // Move highlight up
@@ -261,23 +264,27 @@ export class DropdownComponent extends KeyboardFocusableComponent<DropdownState>
             return;
         }
 
+        // Unfortunately there's a bunch of code that needs to know this
+        // information *later*, so we have to store it
+        this.useMobilePopup = DropdownPositioner.useMobileMode();
+
         // Reset the highlighted option
         this._updatePopupOptionsElement();
 
+        // Configure the popup for desktop/mobile
+        if (this.useMobilePopup) {
+            this.inputBox.inputElement.readOnly = true;
+            this.popupManager.modal = true;
+        } else {
+            this.inputBox.inputElement.readOnly = false;
+            this.popupManager.modal = false;
+        }
+
         // Tell the popup manager to show the popup
         //
-        // This must happen before the options are updated, because the popup
+        // This must happen after the options are updated, because the popup
         // manager needs knowledge about the size of the options.
         this.popupManager.isOpen = true;
-
-        // Make sure mobile browsers don't display a keyboard.
-        //
-        // The popup positioner communicates the layout it has decided on via
-        // CSS classes.
-        this.inputBox.inputElement.readOnly =
-            this.popupElement.classList.contains(
-                "rio-dropdown-popup-mobile-fullscreen"
-            );
     }
 
     /// Close the popup and apply the selected option. Does nothing if the popup
@@ -434,15 +441,13 @@ export class DropdownComponent extends KeyboardFocusableComponent<DropdownState>
                 this._highlightOption(match);
             });
 
-            // With a `click` handler, the <input> element loses focus for a
-            // little while, which is noticeable because the floating label will
-            // quickly move down and then back up. To avoid this, we use
-            // `pointerdown` instead. But on mobile this prevents scrolling, so
-            // mobile uses `click`.
-            match.addEventListener(SELECT_OPTION_EVENT, (event: Event) => {
-                this.hidePopupAndCommit(optionName);
-                markEventAsHandled(event);
-            });
+            match.addEventListener(
+                this.getOptionSelectEvent(),
+                (event: Event) => {
+                    this.hidePopupAndCommit(optionName);
+                    markEventAsHandled(event);
+                }
+            );
         }
 
         // If only one option was found, highlight it