+++ rio/components/rectangle.py
@@ -7,12 +7,30 @@ from uniserde import JsonDoc
 
 import rio
 
-from .. import cursor_style, fills
+from .. import cursor_style, deprecations, fills
 from ..color import Color
 from .fundamental_component import FundamentalComponent
 
 __all__ = [
     "Rectangle",
+]
+
+
+CursorStyle = t.Literal[
+    "default",
+    "none",
+    "help",
+    "pointer",
+    "loading",  # "wait" in CSS
+    "background-loading",  # "progress" in CSS
+    "crosshair",
+    "text",
+    "move",
+    "not-allowed",
+    "can-grab",  # "grab" in CSS
+    "grabbed",  # "grabbing" in CSS
+    "zoom-in",
+    "zoom-out",
 ]
 
 
@@ -148,7 +166,7 @@ class Rectangle(FundamentalComponent):
     _: dataclasses.KW_ONLY
     content: rio.Component | None = None
     transition_time: float = 1.0
-    cursor: rio.CursorStyle = cursor_style.CursorStyle.DEFAULT
+    cursor: CursorStyle | cursor_style.CursorStyle = "default"
     ripple: bool = False
 
     fill: fills._FillLike = Color.TRANSPARENT
@@ -168,6 +186,12 @@ class Rectangle(FundamentalComponent):
     hover_shadow_offset_x: float | None = None
     hover_shadow_offset_y: float | None = None
     hover_shadow_color: rio.Color | None = None
+
+    def __post_init__(self):
+        if isinstance(self.cursor, cursor_style.CursorStyle):
+            deprecations.warn(
+                "`rio.CursorStyle` is deprecated in favor of string literals"
+            )
 
     def _custom_serialize_(self) -> JsonDoc:
         # Impute default values